Move cache lock timeout logic to CacheResultInterceptor
